Rename index.js to BaseCache.js
